requirements.txt
io
openai
streamlit
av==9.2.0
numpy==1.23.0
pydub==0.25.1
streamlit==1.10.0
streamlit_webrtc==0.43.0
azure.cognitiveservices.speech